src/copilot/drawcopilot/polymer/AML_SDFinetune/requirements.txt
adal==1.2.7
alembic==1.11.1
antlr4-python3-runtime==4.8
applicationinsights==0.11.10
argcomplete==2.1.2
atomicwrites==1.4.1
attrs==23.1.0
azure-cli-core==2.49.0
azure-cli-telemetry==1.0.8
azure-common==1.1.28
azure-core==1.27.0
azure-graphrbac==0.61.1
azure-identity==1.13.0
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.2.2
azure-mgmt-resource==22.0.0
azure-mgmt-storage==21.0.0
azure-ml-component==0.9.18
azure-storage-blob==12.13.0
azureml-core==1.51.0
azureml-dataprep==4.10.8
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.17.12
azureml-dataset-runtime==1.51.0
azureml-mlflow==1.51.0
azureml-pipeline-core==1.51.0
azureml-telemetry==1.51.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
blinker==1.6.2
bytecode==0.14.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.0.7
coolname==1.1.0
cryptography==40.0.2
cycler==0.11.0
databricks-cli==0.17.7
decorator==5.1.1
distro==1.8.0
docker==6.1.3
dotnetcore2==3.1.23
entrypoints==0.4
filelock==3.12.0
Flask==2.3.2
flatbuffers==23.5.26
flatten-dict==0.4.2
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.19.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
h5py==3.8.0
humanfriendly==10.0
hydra-core==1.0.4
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpath-ng==1.5.3
jsonpickle==3.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
knack==0.10.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mlflow==2.4.0
mlflow-skinny==2.4.0
mpmath==1.3.0
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
ndg-httpsclient==0.5.1
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.1.2
onnxruntime==1.15.0
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.0.7
packaging==23.0
pandas==2.0.2
paramiko==3.2.0
pathspec==0.11.1
Pillow==9.5.0
pkginfo==1.9.6
pluggy==1.0.0
ply==3.11
portalocker==2.7.0
protobuf==4.23.2
psutil==5.9.5
py==1.11.0
pyarrow==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydash==7.0.4
Pygments==2.15.1
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
querystring-parser==1.2.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.3
shrike==1.34.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.1.0
toml==0.10.2
toposort==1.10
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
waitress==2.1.2
websocket-client==1.5.3
Werkzeug==2.3.6
zipp==3.15.0